+++ tests/Filter/IncludeWeekdayFilterTest.php
@@ -58,7 +58,7 @@ final class IncludeWeekdayFilterTest extends HolidayTestCase
     private function thenACorrectlyFilteredResultShouldBeReturned(array $expectedResult): void
     {
         $resultDates = [];
-        foreach ($this->actualResult->getList() as $result) {
+        foreach ($this->actualResult as $result) {
             $resultDates[] = $result->getSimpleDate();
         }
         self::assertEquals($expectedResult, $resultDates);