+++ tests/Helper/GetHolidaysByNameTest.php
@@ -116,7 +116,7 @@ final class GetHolidaysByNameTest extends HolidayTestCase
     private function thenItShouldReturnAListOfHolidays(array $expectedResult): void
     {
         $actualResult = [];
-        foreach ($this->actualResult->getList() as $holiday) {
+        foreach ($this->actualResult as $holiday) {
             $actualResult[] = $holiday->getSimpleDate();
         }
         self::assertEquals($expectedResult, $actualResult);